requirements.txt
flask==3.0.2
flask-cors==4.0.0
huggingface-hub==0.20.3
python-dotenv==1.0.0
transformers
torchvision 
pillow
bitsandbytes
accelerate>=0.26.0
